maemp/requirements.txt
django==3.1
factory-boy==3.2.0
coverage==5.5
pytest-cov
pytest-django
pytest
psycopg2-binary>=2.7,<3.0